starting on the refactor by creating functions for each util
